pyproject.toml
[tool.poetry]
name = "autogluon-gpu"
version = "0.1.0"
description = ""
authors = ["Hajar Razip <mshajarrazip@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.9"
torch = {url="https://download.pytorch.org/whl/cu113/torch-1.12.1%2Bcu113-cp39-cp39-linux_x86_64.whl"}
autogluon-tabular = {extras = ["all"], version = "~0.6.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
